requirements-hardware.txt
# Optional: only for scripts/serial_bridge.py (laptop + USB Arduino)
pyserial==3.5
python-dotenv==1.1.1
